Refactor standard_two tests
