+++ test/test/standard_two/test_default.py
@@ -13,54 +13,6 @@ def test_default(file = __file__):
             ),
         )
 
-def test_dev_not_found(file = __file__):
-    with TestDirContext(file) as ctx:
-        dockerc = ctx.run_dockerc(
-            'dev',
-        )
-        dockerc.assert_context_ok(
-            format_dockerc_stdout(
-                b'docker compose'
-                b' -f ./docker-compose.yml'
-                b' -f ./docker-compose.override.yml'
-                b' up -d'
-            ),
-        )
-
-def test_override_not_found(file = __file__):
-    with TestDirContext(file) as ctx:
-        dockerc = ctx.run_dockerc(
-            'override',
-        )
-        dockerc.assert_context_ok(
-            format_dockerc_stdout(
-                b'docker compose'
-                b' -f ./docker-compose.yml'
-                b' -f ./docker-compose.override.yml'
-                b' up -d'
-            ),
-        )
-
-def test_prod_not_found(file = __file__):
-    with TestDirContext(file) as ctx:
-        dockerc = ctx.run_dockerc(
-            'prod',
-        )
-        dockerc.assert_context_ok(
-            format_dockerc_stdout(
-                b'docker compose'
-                b' -f ./docker-compose.yml'
-                b' up -d'
-            ),
-        )
-
-def test_what_not_found(file = __file__):
-    with TestDirContext(file) as ctx:
-        dockerc = ctx.run_dockerc(
-            'what',
-        )
-        dockerc.assert_context_not_found()
-
 # def test_default_env(dir_files = dir_files):
 #     reset_dir('./twd', dir_files + [
 #         '.env',
