+++ test/test/standard_two/test_dev.py
@@ -1,0 +1,16 @@
+from test.src.format_dockerc_stdout import format_dockerc_stdout
+from test.src.TestDirContext import TestDirContext
+
+def test_dev_not_found(file = __file__):
+    with TestDirContext(file) as ctx:
+        dockerc = ctx.run_dockerc(
+            'dev',
+        )
+        dockerc.assert_context_ok(
+            format_dockerc_stdout(
+                b'docker compose'
+                b' -f ./docker-compose.yml'
+                b' -f ./docker-compose.override.yml'
+                b' up -d'
+            ),
+        )
